Simplified /core directory structure and renamed a few /core files.
