+++ example/lib/spikes/editor_abstractions/core/document.dart
