+++ example/lib/spikes/editor_abstractions/core/document_editor.dart
@@ -1,13 +1,6 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
 
-// TODO: get rid of these imports
-import '../../default_editor/text.dart';
-import '../../default_editor/box_component.dart';
-import '../../selectable_text/attributed_text.dart';
-import '../../default_editor/paragraph.dart';
-import '../selection/editor_selection.dart';
-import 'rich_text_document.dart';
+import 'document.dart';
 
 class DocumentEditor {
   DocumentEditor({
