+++ example/lib/spikes/editor_abstractions/default_editor/multi_node_editing.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
-import '../core/document/rich_text_document.dart';
-import '../core/document/document_editor.dart';
-import '../core/selection/editor_selection.dart';
+import '../core/document.dart';
+import '../core/document_editor.dart';
+import '../core/document_selection.dart';
 import '../selectable_text/attributed_text.dart';
 import 'text.dart';
 import 'box_component.dart';
