+++ example/lib/spikes/editor_abstractions/default_editor/paragraph.dart
@@ -3,10 +3,10 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 
-import '../core/document/document_editor.dart';
-import '../core/composition/document_composer.dart';
-import '../core/document/rich_text_document.dart';
-import '../core/selection/editor_selection.dart';
+import '../core/document_editor.dart';
+import '../core/document_composer.dart';
+import '../core/document.dart';
+import '../core/document_selection.dart';
 import '../selectable_text/attributed_text.dart';
 import '_text_tools.dart';
 import 'text.dart';
