+++ example/lib/spikes/editor_abstractions/editor.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart' hide SelectableText;
 
-import 'core/document/rich_text_document.dart';
-import 'core/document/document_editor.dart';
+import 'core/document.dart';
+import 'core/document_editor.dart';
 import 'editable_document.dart';
 
 /// Editor for a `RichTextDocument`.
